updated product model and interface in order service

+++ orders/src/app/interfaces/product_interface.py
@@ -7,7 +7,7 @@ class GeneralProductInterface(BaseModel):
     Id: int
     UserId: int
     Name: str
-    Price: int
+    Price: float
     ShortDesc: str
     CreatedAt: datetime
     UpdatedAt: datetime
